mod(Observation_Log): Restructure of observation data flow
- Implementation of BasicDataStore
- Usage if Publisher interface instead of Observations->Inserts
- Observations are parameterizable now

+++ OpenPass_Source_Code/openPASS/CoreModules/EventDetector/CollisionDetector.h
@@ -202,7 +202,7 @@ public:
                                      double &agentDistanceFromInitial,
                                      Common::Vector2d &agentReferencePoint,
                                      Common::Vector2d &otherReferencePoint,
-                                     int &agentPenetratingCorner);
+                                     size_t &agentPenetratingCorner);
 
     //-----------------------------------------------------------------------------
     /*! Calculates parameters of point of contact
